Take crack away and make nick list always available and correct
The @selected_nick_list has been taken out as it's use with the actual nick_list wasn't really working. Now it works on the assumption that the end of the array was the last person to make tea, which is probably safe enough.
Also the nick list exists even before someone calls the first tea although I've just realised that if the cache has expried you'll still get reshuffled.